Move get_syntabledf into extract.py
